pyproject.toml
[tool.poetry]
name = "vending"
version = "0.1.0"
description = ""
authors = ["The Gang Technology"]

[tool.poetry.dependencies]
python = ">=3.7,<4.0"
flask = "^2.2.2"
pyyaml = "^6.0"
python-dotenv = "^0.21.1"
flask-mysqldb = "^1.0.1"
pytest = "^7.2.1"


[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
flake8 = "^5.0.1"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
ipykernel = "^6.15.1"
flake8-annotations = "^2.9.1"
flake8-eradicate = "^1.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
